+++ data-raw/mdt-dict.R
@@ -13,5 +13,11 @@ stopifnot(!anyDuplicated(input$key),
           all(input$key == italian$key))
 input$IT <- italian$IT
 
+german_formal <- read.csv("data-raw/dict-german-formal.csv", stringsAsFactors = FALSE)
+names(german_formal)[[1]] <- "key"
+stopifnot(!anyDuplicated(input$key),
+          all(input$key == german_formal$key))
+input$DE_F <- german_formal$DE_F
+
 mdt_dict <- psychTestR::i18n_dict$new(input)
 usethis::use_data(mdt_dict, overwrite = TRUE)